requirements.txt
numpy
matplotlib
tqdm
Pillow